#350 Split vtm/desktopio
